Changes doen for mercahnt and feeplan

+++ src/app/theme/pages/default/features/merchant/merchant-list/merchant-list.component.ts
@@ -21,7 +21,7 @@ export class MerchantListComponent implements OnInit {
     schoolList: SelectItem[];
     listDisable: boolean;
     schoolId: string;
-    merchantList: Observable<Merchant[]>;
+    merchantList: Merchant[];
     constructor(
         private globalErrorHandler: GlobalErrorHandler,
         private messageService: MessageService,
@@ -61,45 +61,66 @@ export class MerchantListComponent implements OnInit {
         this.merchantServics.getAllMerchants(this.schoolId)
             .subscribe(response => {
                 this.merchantList = response;
+                if (this.merchantList.length == 0) {
+                    this.merchantList.push(new Merchant())
+                }
             });
     }
-
+    addRow() {
+        this.merchantList.push(new Merchant())
+    }
+    enableEdit(row) {
+        row.enable = true;
+    }
     onEditMerchant(row) {
-        this.merchantServics.updateMerchant(row).subscribe(
-            data => {
-                this.messageService.addMessage({ severity: 'success', summary: 'Success', detail: 'Record Deleted Successfully' });
-                this.getAllMerchants();
-            }, error => {
-                this.globalErrorHandler.handleError(error);
-            });
+        row.schoolId= this.schoolId;
+        if (!row.id) {
+            row.id=null;
+            this.merchantServics.createMerchant(row).subscribe(
+                data => {
+                    this.messageService.addMessage({ severity: 'success', summary: 'Success', detail: 'Record Addes Successfully' });
+                    this.getAllMerchants();
+
+                }, error => {
+                    this.globalErrorHandler.handleError(error);
+                });
+        }
+        else {
+            this.merchantServics.updateMerchant(row).subscribe(
+                data => {
+                    this.messageService.addMessage({ severity: 'success', summary: 'Success', detail: 'Record Updated Successfully' });
+                    this.getAllMerchants();
+                }, error => {
+                    this.globalErrorHandler.handleError(error);
+                });
+        }
     }
-
-    onDeleteMerchant(id) {
-        this.confirmationService.confirm({
-            message: 'Do you want to delete this record?',
-            header: 'Delete Confirmation',
-            icon: 'fa fa-trash',
-            accept: () => {
-                this.merchantServics.deleteMerchant(id).subscribe(
-                    data => {
-                        this.messageService.addMessage({ severity: 'success', summary: 'Success', detail: 'Record Deleted Successfully' });
-                    }, error => {
-                        this.globalErrorHandler.handleError(error);
-                    });
-            },
-            reject: () => {
-            }
-        });
-        // this.merchantServics.deleteMerchant(id)
-        // .subscribe(response => {
-
-        // });
-        this.getAllMerchants();
+    onDeleteMerchant(id, index) {
+        if (id) {
+            this.confirmationService.confirm({
+                message: 'Do you want to delete this record?',
+                header: 'Delete Confirmation',
+                icon: 'fa fa-trash',
+                accept: () => {
+                    this.merchantServics.deleteMerchant(id).subscribe(
+                        data => {
+                            this.messageService.addMessage({ severity: 'success', summary: 'Success', detail: 'Record Deleted Successfully' });
+                            this.getAllMerchants();
+                        }, error => {
+                            this.globalErrorHandler.handleError(error);
+                        });
+                },
+                reject: () => {
+                }
+            });
+        }
+        else {
+            this.merchantList.splice(index, 1);
+        }
     }
 
     onSchoolSelect(id) {
         this.schoolId = id;
-        this.listDisable = true;
         this.getAllMerchants();
         // if (!localStorage.getItem("schoolId") || localStorage.getItem("schoolId") == "null" || localStorage.getItem("schoolId") == "0") {
         //     this.messageService.addMessage({ severity: 'error', summary: 'Error', detail: 'Please Select School' });
